+++ example_systems/1_three_zones_asymmetric_bidirectional_Actual_Symmetric_2nd copy/Run.jl
@@ -1,0 +1,3 @@
+using GenX
+
+run_genx_case!(dirname(@__FILE__))