+++ src/configure_settings/configure_settings.jl
@@ -37,7 +37,9 @@ function default_settings()
         "ResourcePoliciesFolder" => "policy_assignments",
         "SystemFolder" => "system",
         "PoliciesFolder" => "policies",
-        "ObjScale" => 1)
+        "ObjScale" => 1,
+        "AsymmetricalTransFlowLimit" => 0
+    )
 end
 
 @doc raw"""