+++ src/model/core/transmission/investment_transmission.jl
@@ -1,16 +1,34 @@
 @doc raw"""
     investment_transmission!(EP::Model, inputs::Dict, setup::Dict)
-This function model transmission expansion and adds transmission reinforcement or construction costs to the objective function. Transmission reinforcement costs are equal to the sum across all lines of the product between the transmission reinforcement/construction cost, $\pi^{TCAP}_{l}$, times the additional transmission capacity variable, $\bigtriangleup\varphi^{cap}_{l}$.
+This function model transmission expansion and adds transmission reinforcement or construction costs to the objective function. Transmission reinforcement costs are equal to the sum across all lines of the product between the transmission reinforcement/construction cost, $\pi^{TCAP}_{l}$, times the additional transmission capacity variable, $\bigtriangleup\varphi^{cap}_{l}$. 
+For asymmetric bidirectional flows, the reinforcement costs are equal to the sum across all lines of the product between the transmission reinforcement/construction costs for the onward and return directions, $\pi^{TCAP+}_{l}$ and $\pi^{TCAP-}_{l}$, respectively, times the additional transmission capacity variables, $\bigtriangleup\varphi^{cap+}_{l}$ and $\bigtriangleup\varphi^{cap-}_{l}$
 ```math
 \begin{aligned}
     & \sum_{l \in \mathcal{L}}\left(\pi^{TCAP}_{l} \times \bigtriangleup\varphi^{cap}_{l}\right)
 \end{aligned}
 ```
+For asymmetric bidirectional lines
+
+```math
+\begin{aligned}
+    & \sum_{l \in \mathcal{L_{asym}}}\left(\pi^{TCAP+}_{l} \times \bigtriangleup\varphi^{cap+}_{l}\right)
+\end{aligned}
+```
+
+and 
+
+```math
+\begin{aligned}
+    & \sum_{l \in \mathcal{L_{asym}}}\left(\pi^{TCAP-}_{l} \times \bigtriangleup\varphi^{cap-}_{l}\right)
+\end{aligned}
+```
+
 Note that fixed O\&M and replacement capital costs (depreciation) for existing transmission capacity is treated as a sunk cost and not included explicitly in the GenX objective function.
 
 **Accounting for Transmission Between Zones**
 
-Available transmission capacity between zones is set equal to the existing line's maximum power transfer capacity, $\overline{\varphi^{cap}_{l}}$, plus any transmission capacity added on that line (for lines eligible for expansion in the set $\mathcal{E}$). 
+Available transmission capacity between zones is set equal to the existing line's maximum power transfer capacity, $\overline{\varphi^{cap}_{l}}$ ($\overline{\varphi^{cap+}_{l}}$ 
+and $\overline{\varphi^{cap-}_{l}}$ for positive and negative directions, respectively, for asymmetrical lines), plus any transmission capacity added on that line (for lines eligible for expansion in the set $\mathcal{E}$). 
 ```math
 \begin{aligned}
     &\varphi^{cap}_{l} = \overline{\varphi^{cap}_{l}} , &\quad \forall l \in (\mathcal{L} \setminus \mathcal{E} ),\forall t  \in \mathcal{T}\\
@@ -18,7 +36,9 @@ Available transmission capacity between zones is set equal to the existing line'
     &\varphi^{cap}_{l} = \overline{\varphi^{cap}_{l}} + \bigtriangleup\varphi^{cap}_{l} , &\quad \forall l \in \mathcal{E},\forall t  \in \mathcal{T}        
 \end{aligned}
 ```
-The additional transmission capacity, $\bigtriangleup\varphi^{cap}_{l} $, is constrained by a maximum allowed reinforcement, $\overline{\bigtriangleup\varphi^{cap}_{l}}$, for each line $l \in \mathcal{E}$.
+The additional transmission capacity, $\bigtriangleup\varphi^{cap}_{l} $ ($\bigtriangleup\varphi^{cap+}_{l} $ and $\bigtriangleup\varphi^{cap-}_{l} $ respectively for asymmetric lines), 
+is constrained by a maximum allowed reinforcement, $\overline{\bigtriangleup\varphi^{cap}_{l}}$ ($\overline{\bigtriangleup\varphi^{cap+}_{l}}$ 
+and $\overline{\bigtriangleup\varphi^{cap-}_{l}}$ respectively for asymmetric lines), for each line $l \in \mathcal{E}$.
 ```math
 \begin{aligned}
     & \bigtriangleup\varphi^{cap}_{l}  \leq \overline{\bigtriangleup\varphi^{cap}_{l}}, &\quad \forall l \in \mathcal{E}
@@ -28,54 +48,81 @@ The additional transmission capacity, $\bigtriangleup\varphi^{cap}_{l} $, is con
 function investment_transmission!(EP::Model, inputs::Dict, setup::Dict)
     println("Investment Transmission Module")
 
-    L = inputs["L"]     # Number of transmission lines
     NetworkExpansion = setup["NetworkExpansion"]
     MultiStage = setup["MultiStage"]
 
+    SYMMETRIC_LINE_INDEX = inputs["SYMMETRIC_LINE_INDEX"]
+    ASYMMETRIC_LINE_INDEX = inputs["ASYMMETRIC_LINE_INDEX"]
+
     if NetworkExpansion == 1
         # Network lines and zones that are expandable have non-negative maximum reinforcement inputs
-        EXPANSION_LINES = inputs["EXPANSION_LINES"]
+        ASYMMETRIC_EXPANSION_LINES = inputs["ASYMMETRIC_EXPANSION_LINES"]
+        SYMMETRIC_EXPANSION_LINES = inputs["SYMMETRIC_EXPANSION_LINES"]
     end
 
     ### Variables ###
 
     if MultiStage == 1
-        @variable(EP, vTRANSMAX[l = 1:L]>=0)
+        @variable(EP, vTRANSMAX[l in SYMMETRIC_LINE_INDEX]>=0)
+        @variable(EP, vTRANSMAX_Pos[l in ASYMMETRIC_LINE_INDEX]>=0)
+        @variable(EP, vTRANSMAX_Neg[l in ASYMMETRIC_LINE_INDEX]>=0)
     end
 
     if NetworkExpansion == 1
         # Transmission network capacity reinforcements per line
-        @variable(EP, vNEW_TRANS_CAP[l in EXPANSION_LINES]>=0)
+        @variable(EP, vNEW_TRANS_CAP[l in SYMMETRIC_EXPANSION_LINES]>=0)
+        @variable(EP, vNEW_TRANS_CAP_Pos[l in ASYMMETRIC_EXPANSION_LINES]>=0)
+        @variable(EP, vNEW_TRANS_CAP_Neg[l in ASYMMETRIC_EXPANSION_LINES]>=0)
     end
 
     ### Expressions ###
 
     if MultiStage == 1
-        @expression(EP, eTransMax[l = 1:L], vTRANSMAX[l])
+        @expression(EP, eTransMax[l in SYMMETRIC_LINE_INDEX], vTRANSMAX[l])
+        @expression(EP, eTransMax_Pos[l in ASYMMETRIC_LINE_INDEX], vTRANSMAX_Pos[l])
+        @expression(EP, eTransMax_Neg[l in ASYMMETRIC_LINE_INDEX], vTRANSMAX_Neg[l])
     else
-        @expression(EP, eTransMax[l = 1:L], inputs["pTrans_Max"][l])
+        @expression(EP, eTransMax[l in SYMMETRIC_LINE_INDEX], inputs["pTrans_Max"][l])
+        @expression(EP, eTransMax_Pos[l in ASYMMETRIC_LINE_INDEX], inputs["pTrans_Max"][l])
+        @expression(EP, eTransMax_Neg[l in ASYMMETRIC_LINE_INDEX], inputs["pTrans_Max_Neg"][l])
     end
 
     ## Transmission power flow and loss related expressions:
     # Total availabile maximum transmission capacity is the sum of existing maximum transmission capacity plus new transmission capacity
     if NetworkExpansion == 1
-        @expression(EP, eAvail_Trans_Cap[l = 1:L],
-            if l in EXPANSION_LINES
+        @expression(EP, eAvail_Trans_Cap[l in SYMMETRIC_LINE_INDEX],
+            if l in SYMMETRIC_EXPANSION_LINES
                 eTransMax[l] + vNEW_TRANS_CAP[l]
             else
                 eTransMax[l]
             end)
+        @expression(EP, eAvail_Trans_Cap_Pos[l in ASYMMETRIC_LINE_INDEX],
+            if l in ASYMMETRIC_EXPANSION_LINES
+                eTransMax_Pos[l] + vNEW_TRANS_CAP_Pos[l]
+            else
+                eTransMax_Pos[l]
+            end)
+        @expression(EP, eAvail_Trans_Cap_Neg[l in ASYMMETRIC_LINE_INDEX],
+            if l in ASYMMETRIC_EXPANSION_LINES
+                eTransMax_Neg[l] + vNEW_TRANS_CAP_Neg[l]
+            else
+                eTransMax_Neg[l]
+            end)
     else
-        @expression(EP, eAvail_Trans_Cap[l = 1:L], eTransMax[l])
+        @expression(EP, eAvail_Trans_Cap[l in SYMMETRIC_LINE_INDEX], eTransMax[l])
+        @expression(EP, eAvail_Trans_Cap_Pos[l in ASYMMETRIC_LINE_INDEX], eTransMax_Pos[l])
+        @expression(EP, eAvail_Trans_Cap_Neg[l in ASYMMETRIC_LINE_INDEX], eTransMax_Neg[l])
     end
-
     ## Objective Function Expressions ##
 
     if NetworkExpansion == 1
         @expression(EP,
             eTotalCNetworkExp,
             sum(vNEW_TRANS_CAP[l] * inputs["pC_Line_Reinforcement"][l]
-            for l in EXPANSION_LINES))
+                for l in SYMMETRIC_EXPANSION_LINES; init = 0) + 
+            sum((vNEW_TRANS_CAP_Pos[l] + vNEW_TRANS_CAP_Neg[l]) * inputs["pC_Line_Reinforcement"][l]
+                for l in ASYMMETRIC_EXPANSION_LINES; init = 0)
+            )
 
         if MultiStage == 1
             # OPEX multiplier to count multiple years between two model stages
@@ -93,7 +140,9 @@ function investment_transmission!(EP::Model, inputs::Dict, setup::Dict)
 
     if MultiStage == 1
         # Linking constraint for existing transmission capacity
-        @constraint(EP, cExistingTransCap[l = 1:L], vTRANSMAX[l]==inputs["pTrans_Max"][l])
+        @constraint(EP, cExistingTransCap[l in SYMMETRIC_LINE_INDEX], vTRANSMAX[l]==inputs["pTrans_Max"][l])
+        @constraint(EP, cExistingTransCapPos[l in ASYMMETRIC_LINE_INDEX], vTRANSMAX_Pos[l]==inputs["pTrans_Max"][l])
+        @constraint(EP, cExistingTransCapNeg[l in ASYMMETRIC_LINE_INDEX], vTRANSMAX_Neg[l]==inputs["pTrans_Max_Neg"][l])
     end
 
     # If network expansion is used:
@@ -102,13 +151,25 @@ function investment_transmission!(EP::Model, inputs::Dict, setup::Dict)
         if MultiStage == 1
             # Constrain maximum possible flow for lines eligible for expansion regardless of previous expansions
             @constraint(EP,
-                cMaxFlowPossible[l in EXPANSION_LINES],
+                cMaxFlowPossible[l in SYMMETRIC_EXPANSION_LINES],
                 eAvail_Trans_Cap[l]<=inputs["pTrans_Max_Possible"][l])
+            @constraint(EP,
+                cMaxFlowPossible_Pos[l in ASYMMETRIC_EXPANSION_LINES],
+                eAvail_Trans_Cap_Pos[l]<=inputs["pTrans_Max_Possible"][l])
+            @constraint(EP,
+                cMaxFlowPossible_Neg[l in ASYMMETRIC_EXPANSION_LINES],
+                eAvail_Trans_Cap_Neg[l]<=inputs["pTrans_Max_Possible_Neg"][l])
         end
         # Constrain maximum single-stage line capacity reinforcement for lines eligible for expansion
         @constraint(EP,
-            cMaxLineReinforcement[l in EXPANSION_LINES],
+            cMaxLineReinforcement[l in SYMMETRIC_EXPANSION_LINES],
             vNEW_TRANS_CAP[l]<=inputs["pMax_Line_Reinforcement"][l])
+        @constraint(EP,
+            cMaxLineReinforcement_Pos[l in ASYMMETRIC_EXPANSION_LINES],
+            vNEW_TRANS_CAP_Pos[l]<=inputs["pMax_Line_Reinforcement"][l])
+        @constraint(EP,
+            cMaxLineReinforcement_Neg[l in ASYMMETRIC_EXPANSION_LINES],
+            vNEW_TRANS_CAP_Neg[l]<=inputs["pMax_Line_Reinforcement_Neg"][l])
     end
     #END network expansion contraints